+++ core/src/main/scala/ch/epfl/scala/debugadapter/internal/scalasig/Ref.scala
@@ -1,4 +1,4 @@
-package ch.epfl.scala.debugadapter.internal.decompiler.scalasig
+package ch.epfl.scala.debugadapter.internal.scalasig
 
 import scala.language.implicitConversions
 import scala.reflect.ClassTag
