+++ core/src/main/scala/ch/epfl/scala/debugadapter/internal/scalasig/ScalaSigPrinter.scala
@@ -5,7 +5,7 @@
 **
  */
 
-package ch.epfl.scala.debugadapter.internal.decompiler.scalasig
+package ch.epfl.scala.debugadapter.internal.scalasig
 
 import java.lang.StringBuilder
 import java.util.regex.Pattern
