+++ core/src/main/scala/ch/epfl/scala/debugadapter/internal/scalasig/ScalaSigReader.scala
@@ -1,4 +1,4 @@
-package ch.epfl.scala.debugadapter.internal.decompiler.scalasig
+package ch.epfl.scala.debugadapter.internal.scalasig
 
 import java.nio.charset.StandardCharsets
